+++ src/mesh/forest/forest.cpp
@@ -26,7 +26,7 @@
 
 #include "basic_types.hpp"
 #include "defs.hpp"
-#include "mesh/forest.hpp"
+#include "mesh/forest/forest.hpp"
 #include "mesh/logical_location.hpp"
 #include "utils/bit_hacks.hpp"
 #include "utils/indexer.hpp"
